Fixed linting
